Change NOTREACHED to NOTREACHED()
It was suggested that using:

        err(...);
        NOTREACHED();

Is better because it makes NOTREACHED look like a function.
On systems where __builtin_unreachable is not implemented
(or even where __has_builtin is not defined, NOTREACHED()
turns into a call to abort(), so making it look like a function,
as was suggested, seems like a good thing to do.