Update dbg_example.c to match dbg.3
Due to more than one person working on this repo and apparently my
updates to dbg.3 being updated here without the dbg_example.c being
updated this commit makes dbg_example.c be the same as in dbg.3 as it
should have been in the first place (this happened as I was having a
problem with my fork).

Now dbg_example.c shows what the next line should be so that the user
can verify that it works properly.

I also added a missing full stop in a functions list in dbg.3.

As well I removed a .RE from dbg.3 that did not have a matching .RS
(and does not need one).