Improve Makefile, build dbg.a
The Makefile has been improved to build the dbg.a library.

Added extern_include extern_objs extern_liba extern_all rules
to make it easier for other Makefiles to trigger the build
of various components.

Improved Makefile comments.

Update .gitignore to explicitly ignore select files including
the new dbg.a library.

The source code was NOT updated and thus the dbg version
was NOT changed.