Add man pages for different types of messages
This means that the following main man pages were added:

    err.3
    warn.3
    werr.3
    warn_or_err.3
    printf_usage.3
    msg.3

and the dbg.3 has been modified to only refer to (in description) the
dbg() specific functions.

Also because a lot of man pages have been added aside from these I moved
the man pages to man/man3 (like in an installed system).

A note is that there is room for improvement here. This is a work in
progress but enough work has been done to commit.

An important point is that although the Makefile now installs those man
pages it does not install the man pages for the additional man pages
with the .so man3/foo.3 in it. This is because there are a lot of them
and the way to go about it might be up to discussion in the mkiocccentry
repo first. There even are some other ideas as opposed to using .so.

Another method might be to use symlinks so that one can do

     man ./man/man3/werr.3

for example and get the man page output whereas this will not happen if
it's not installed. Another option is to have make install use ln to
make symlinks. The way it currently stands is that one can only do:

    man man/man3/err.3
    man man/man3/warn.3
    man man/man3/werr.3
    man man/man3/warn_or_err.3
    man man/man3/printf_usage.3
    man man/man3/msg.3

and expect the right output as each one will have the group of
functions. But whether we want to make it symlinks instead - that might
be an idea. Depending on what is decided will change how make install
works.

The groups I selected could be changed as well but this is a good start.