release dbg 2.10 2023-08-01
Improved checks for when a file stream is NULL, or
when stderr is NULL.  On both cases, the function
avoid dereferencing a NULL FILE *.

Fixed a few places where a message was sent
to stderr when it should have been written
to the stream argument.

Updated README.md by adding the usual TL;DR
preamble for how to compile, test and install dbg.
Added a few notes about fairly robust arg checking.