Add parse_verbosity(3) to API
Since the debug facility is the one that uses a verbosity_level I have
added the function that I added to the mkiocccentry repo util.c file
(under jparse/util.c) to the dbg.c API.

The dbg.3 man page has been updated with this and also some other
improvements. There is a recursive reference to the man page for
parse_verbosity(3) in the man page just for fun.

The dbg related changes will be added to the mkiocccentry repo so that
these changes are synchronised to that repo.

The dbg_example.c file that I wrote for this repo and the mkiocccentry
repo (initially) has also been updated with a comment change, referring
to not strtol(3) but rather the parse_verbosity(3) function.