Rename dbg.a libdbg.a
To allow for linking in this library from say /usr/local/lib the library
file has to be renamed to /usr/local/libdbg.a. This is a linker
implementation detail. We need to have the ability to link in the dbg
API in the jparse repo (https://github.com/xexyl/jparse) and this now is
easier to link in and thus use the library.

The README.md set up instructions have two ways now: first what already
existed, the set up of the library if you do not wish to install it and
now the second, how to do it if you wish to install it.

For jparse another library needs this change which will be done after
this, though possibly not right away.