Use -L. -ldbg in Makefile
Instead of directly linking in libdbg.a use -L. -ldbg in the Makefile,
so it is more like a library (which it is). One may still use the other
way (linking it in as if it is an object file) if they so wish but this
might be better for example purposes too.