Improve Makefile (install, uninstall)
Add to the Makefile the PREFIX variable so one can change the install
location from the default /usr/local. For instance to use /usr:

    make PREFIX=/usr install

Added uninstall rule to uninstall the programs, library and man pages.
If one specified a different PREFIX in install they would have to
specify it for uninstall too like:

    make PREFIX=/usr uninstall

To be congruent with install, the uninstall allows one to silence the rm
verbosity (RM_V variable instead of INSTALL_V variable).

Updated README.md file for the above changes (well not the RM_V part).