Add timestamps to starting/ending in Makefile
A question of whether to have it in the clobber/clean related rules is
something to consider but since the rules do have starting/ending
messages already and since they do not recurse into any subdirectory it
does not seem to be a problem. This of course can change if necessary.