Fix rm in Makefile rules
Add where missing ${Q} to ${RM} lines.

Add where it was not ${Q} ${Q} at the beginning of ${RM} lines.

Add the ${RM_V} to the ${RM} lines where it was missing (it is
analogous to ${INSTALL_V} but the default of -v could certainly be
changed if it is a problem).