Clarify slightly a comment in dbg_example.c
I should have put () after main to make it clearer that the function to
check is not the main function of the API (in the sense of the main
function of the file which could be interpreted as dbg(), unlikely as
that seems) but the actual main() function instead.

This was updated in the dbg.3 man page too as that example is there (it
is useful to have it in the man page since make install does not install
dbg_example.c but it does install the man pages).