Improve Makefile, add verbosity to make test
Rename Q_V_OPTION to 'VERBOSITY' in Makefile for easier use.

Update GitHub workflow test file to use make VERBOSITY=7 test. The
dbg_test.c uses fdbg() (on stderr) but the make test rule writes to an
output file that then is validated so the make test with verbosity does
not do much currently but it allows for one to do more if is desired.