Disable -g3 in Makefile
The use of -O3 with debug symbols is problematic because by optimising
the code it makes it harder to debug even with debug symbols generated
(-g option). If one needs to have debug symbols they should do something
like:

    make C_OPT=-g3

and not have the -O3 (they could use -O0 instead of course).