add is_dbg_enabled macro, improve dbg_example.c, update dbg.3
We add a new is_dbg_enabled macro based on a call to `dbg_allowed(1)`.

We improved some of the examples in `dbg_example.c`.

We updated and improved the `/man/man3/dbg.3` man page.

Changed DBG_VERSION from "3.0 2023-08-05"
to "3.1 2023-08-28".

Ran `make clobber all test` to perform and test the above.