improve man page SYNOPSIS
Adopt section 3 style `SYNOPSIS` for functions: removed `extern`,
start with a return type preceding line, and a blank line between
functions.

Removed `__attribute__` text on functions in `SYNOPSIS` as this
is overly complex and doesn't assist the reader that much.

Fixed dates on man pages to the date of this commit.

Align comments in the `SYNOPSIS` section, and remove problematic
ASCII tab characters that complicate man page formatting.

Fixed `is_dbg_enabled` in `dbg(3)` man page as this is a macro.